+++ scripts/macros/2014/classFeatures/cleric/natureDomain/charmAnimalsAndPlants.js
@@ -1,8 +1,7 @@
-import {actorUtils, genericUtils} from '../../../../../utils.js';
-
+import {actorUtils, workflowUtils} from '../../../../../utils.js';
 async function early({workflow}) {
     let validTargets = workflow.targets.filter(i => ['beast', 'plant'].includes(actorUtils.typeOrRace(i.actor)));
-    genericUtils.updateTargets(validTargets);
+    await workflowUtils.updateTargets(workflow, validTargets);
 }
 export let charmAnimalsAndPlants = {
     name: 'Channel Divinity: Charm Animals and Plants',