+++ scripts/macros/2014/spells/thunderStep.js
@@ -1,6 +1,5 @@
 import {Teleport} from '../../../lib/teleport.js';
-import {actorUtils, animationUtils, dialogUtils, genericUtils, itemUtils, tokenUtils} from '../../../utils.js';
-
+import {actorUtils, animationUtils, dialogUtils, genericUtils, itemUtils, tokenUtils, workflowUtils} from '../../../utils.js';
 async function early({workflow}) {
     let casterSize = actorUtils.getSize(workflow.actor);
     let nearbyTargets = tokenUtils.findNearby(workflow.token, 5, 'ally', {includeIncapacitated: true}).filter(i => actorUtils.getSize(i.actor) <= casterSize);
@@ -11,7 +10,7 @@ async function early({workflow}) {
             toTeleport.push(selection[0]);
         }
     }
-    genericUtils.updateTargets(workflow.targets.filter(i => !toTeleport.includes(i)));
+    await workflowUtils.updateTargets(workflow, Array.from(workflow.targets.filter(i => !toTeleport.includes(i))));
     let playAnimation = itemUtils.getConfig(workflow.item, 'playAnimation') && animationUtils.jb2aCheck();
     await Teleport.group(toTeleport, workflow.token, {range: 90, animation: playAnimation ? 'thunderStep' : 'none'});
 }