+++ scripts/macros/2024/classFeatures/rogue/assassin/infiltrationExpertise.js
@@ -1,0 +1,5 @@
+export let infiltrationExpertise = {
+    name: 'Infiltration Expertise',
+    version: '1.3.53',
+    rules: 'modern'
+};